tack max function back where shuld be
